+++ src/app/op-log/core/types/apply.types.ts
@@ -1,2 +1,23 @@
-// Re-exported from @sp/sync-core. Source of truth lives in packages/sync-core/src/apply.types.ts.
-export type { ApplyOperationsResult, ApplyOperationsOptions } from '@sp/sync-core';
+// App-narrowed apply types. The lib's @sp/sync-core ships generic versions; the
+// app uses its own Operation type so callers see the SP-narrowed entityType /
+// actionType unions and the syncImportReason field.
+
+import type { Operation } from '../operation.types';
+
+export interface ApplyOperationsResult {
+  /** Operations that were successfully applied. */
+  appliedOps: Operation[];
+  failedOp?: {
+    op: Operation;
+    error: Error;
+  };
+}
+
+export interface ApplyOperationsOptions {
+  /**
+   * When true, skip side effects that would normally fire on first application
+   * (e.g. writing to archive storage) — used when replaying already-persisted
+   * local operations during hydration.
+   */
+  isLocalHydration?: boolean;
+}
